+++ pages/view.php
@@ -7,7 +7,7 @@
  */
 include_once "../include/db.php";
 
-$ref=getval("ref",0,true);
+$ref=(int) getval("ref",0,true);
 
 # External access support (authenticate only if no key provided, or if invalid access key provided)
 $k=getval("k","");if (($k=="") || (!check_access_key($ref,$k))) {include "../include/authenticate.php";}
